Wayland: Clean up dependencies on gen headers
In prior versions of Android, we needed to emit an explicit dependency
for a special export_includes file (otherwise generated internally to
the Android build system), so that the header files would be generated
before any code outside the library could incldue them.

This dependency was not correct for x86-64 code, as there were two
such export_include files, one under $OUT/obj/..., the other under
$OUT/obj_64/..., and only the first had the headers as a source
dependency.

With the N branch, the Android build system does now have a way to
correctly specify that the generated headers should be a dependency --
they can be added to LOCAL_GENERATED_SOURCES and now all files in that
list will be listed as a dependency, as well as any implementation files
in the same list also being compiled and included in the library.

This patch removes the lines which set up the additional dependencies on
the headers, as that really is an internal detail of the build system,
and now sets LOCAL_GENERATED_SOURCES to include headers as well as
implementation files.

Bug: 65008174
Test: Clean build x86-64
Test: Incremental build with generated files deleted first
Change-Id: Ieb12144401f186e5fedb02fb91c97c722a0d167c