Add host shared library target
Add generated versions of the protocol as 'public-code' instead of
'private-code' so that various symbols are exported from the shared
library and can be used downstream by e.g. crosvm.

This change includes a workaround to suppress the generation of 32-bit
x86 host shared libraries, as these are currently broken due to libffi
not compiling correctly for 32-bit x86. We don't need them and thus can
safely disable.

Due to the split between shared and static, this change also renames the
static version to "_static" so that the host shared library on disk has
the same name as upstream.

Change-Id: I6e6b0f154ce9e45830ed42d23847b23e0761ddc5